# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Provides validator block rewards and sync duties information.",
        "title": "Ethereum Validator API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Tsvetan Dimitrov",
            "email": "tsvetan.dimitrov23@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/blockreward/{slot}": {
            "get": {
                "description": "Retrieves block reward details for a given slot.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BlockReward"
                ],
                "summary": "Get Block Reward",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Slot number",
                        "name": "slot",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.blockRewardResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.APIError"
                        }
                    }
                }
            }
        },
        "/syncduties/{slot}": {
            "get": {
                "description": "Retrieves validators assigned for sync committee duties for a given slot.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SyncDuties"
                ],
                "summary": "Get Sync Duties",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Slot number",
                        "name": "slot",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.syncDutiesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.APIError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "details": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "handlers.blockRewardResponse": {
            "type": "object",
            "properties": {
                "reward": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "handlers.syncDutiesResponse": {
            "type": "object",
            "properties": {
                "validators": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        }
    }
}